Move environment into knowledge
