+++ scooby/__init__.py
@@ -15,7 +15,8 @@ Werthmüller for ``empymod``, ``emg3d``, and the ``SimPEG`` framework
 ``watermark.py`` from https://github.com/rasbt/watermark.
 """
 
-from scooby.report import Report, get_version, in_ipython, in_ipykernel
+from scooby.report import Report, get_version
+from scooby.knowledge import in_ipython, in_ipykernel
 
 __all__ = ['Report', 'in_ipython', 'in_ipykernel', 'get_version']
