+++ scooby/report.py
@@ -7,6 +7,7 @@ import multiprocessing
 from types import ModuleType
 
 from .knowledge import VERSION_ATTRIBUTES, VERSION_METHODS, MKL_INFO, TOTAL_RAM
+from .knowledge import in_ipython, in_ipykernel
 
 
 # Info classes
@@ -283,7 +284,7 @@ class Report(PlatformInfo, PythonInfo):
         return html
 
 
-# Functionalities that might also be of interest on its own.
+# This functionaliy might also be of interest on its own.
 def get_version(module, optional):
     """Get the version of `module` by passing the package or it's name.
 
@@ -357,41 +358,3 @@ def get_version(module, optional):
 
         # If not found, return 'Version unknown'
         return name, 'Version unknown'
-
-
-def in_ipython():
-    """Mystery: are we in an IPython environment?
-
-    Returns
-    -------
-    bool : True if in an IPython environment
-    """
-    try:
-        __IPYTHON__
-        return True
-    except NameError:
-        return False
-
-
-def in_ipykernel():
-    """Mystery: are we in a ipykernel (most likely Jupyter) environment?
-
-    Warning
-    -------
-    There is no way to tell if the code is being executed in a notebook
-    (Jupyter Notebook or Jupyter Lab) or a kernel is used but executed in a
-    QtConsole, or in an IPython console, or any other frontend GUI. However, if
-    `in_ipykernel` returns True, you are most likely in a Jupyter Notebook/Lab,
-    just keep it in mind that there are other possibilities.
-
-    Returns
-    -------
-    bool : True if using an ipykernel
-    """
-    ipykernel = False
-    if in_ipython():
-        try:
-            ipykernel = type(get_ipython()).__module__.startswith('ipykernel.')
-        except NameError:
-            pass
-    return ipykernel
